requirements.txt
backcall==0.1.0
certifi==2019.6.16
chardet==3.0.4
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.6
django-allauth==0.39.1
django-environ==0.4.5
django-extensions==2.2.1
django-recaptcha==2.0.4
gunicorn==19.9.0
idna==2.8
ipython==7.7.0
ipython-genutils==0.2.0
jedi==0.14.1
oauthlib==3.0.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
psycopg2==2.8.3
psycopg2-binary==2.8.3
ptyprocess==0.6.0
py3dns==3.2.0
Pygments==2.4.2
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
stripe==2.37.2
traitlets==4.3.2
urllib3==1.25.6
validate-email==1.3
wcwidth==0.1.7
whitenoise==4.1.4
